requirements.txt
scapy>=2.5.0
python-nmap>=0.7.1
paramiko>=3.0.0
ldap3>=2.9.1
